Enable parsing headers with config and uninitialized buffer
This exposes a new `Headers` type and the existing `HeaderParserConfig`
type in order to enable parsing headers with a custom config or with an
uninitialized buffer.

The `Headers` type enables using uninitialized buffers in a manner
consistent with `Request` and `Response`. An alternative might be to
avoid using `Headers` altogether and have an interface that accepts both
the uninitialized headers and a `&mut &[Header]` where a pointer to the
initialized headers could be written, but that's somewhat ugly and
inconsistent with the `Request` and `Response` interfaces.